split base class into screen and experiment
