+++ src/pymovements/datasets/gazebase.py
@@ -23,7 +23,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from pymovements.datasets.public_dataset import PublicDataset
-from pymovements.gaze.base import Experiment
+from pymovements.gaze.experiment import Experiment
 
 
 class GazeBase(PublicDataset):
