+++ src/pymovements/datasets/public_dataset.py
@@ -27,7 +27,7 @@ from typing import Any
 from urllib.error import URLError
 
 from pymovements.datasets.dataset import Dataset
-from pymovements.gaze.base import Experiment
+from pymovements.gaze.experiment import Experiment
 from pymovements.utils.archives import extract_archive
 from pymovements.utils.downloads import download_file
