+++ src/pymovements/gaze/__init__.py
@@ -19,9 +19,14 @@
 # SOFTWARE.
 """This module holds gaze time series related functionality."""
 from pymovements.gaze import transforms  # noqa: F401
+from pymovements.gaze.experiment import Experiment
 from pymovements.gaze.gaze_dataframe import GazeDataFrame  # noqa: F401
+from pymovements.gaze.screen import Screen
+
 
 __all__ = [
+    'Experiment',
     'GazeDataFrame',
+    'Screen',
     'transforms',
 ]
